create Loadable mixin
